+++ src/api/routers/static_router.py
