+++ src/api/terminal_service.py
